+++ compat/ffmpeg-peaks/index.js
@@ -1,1 +1,2 @@
-export { default } from "peakwright";
+export * from "@peakwright/node";
+export { default } from "@peakwright/node";